NP-532 Split Worklist into multiple files
