+++ src/AppRoutes.tsx
@@ -12,7 +12,7 @@ import { RootStore } from './redux/reducers/rootReducer';
 import { checkIfAppAdmin, checkIfPublisher, checkIfCurator } from './utils/authorization';
 import PublicProfile from './pages/publication/PublicProfile';
 import InstututionPage from './pages/InstitutionsPage';
-import Worklist from './pages/Worklist';
+import WorklistPage from './pages/worklist/WorklistPage';
 
 const AppRoutes: FC = () => {
   const user = useSelector((store: RootStore) => store.user);
@@ -26,7 +26,7 @@ const AppRoutes: FC = () => {
       {isPublisher && <Route exact path="/new-publication" component={NewPublication} />}
       {isPublisher && <Route exact path="/my-publications" component={MyPublications} />}
       {isAppAdmin && <Route exact path="/admin-institutions" component={InstututionPage} />}
-      {isCurator && <Route exact path="/worklist" component={Worklist} />}
+      {isCurator && <Route exact path="/worklist" component={WorklistPage} />}
       {user.isLoggedIn && <Route exact path="/public-profile/:userName" component={PublicProfile} />}
       <Route exact path="/search" component={Search} />
       <Route exact path="/publication/:publicationId" component={PublicationPage} />
